refactor: move tasks and users to dedicated pages
